pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openvocab-eval"
version = "0.1.0"
description = "Open-vocabulary evaluation with CLIP + (optional) LLM-driven attribute prompting."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Your Name"}]
keywords = ["CLIP", "open-vocabulary", "evaluation", "prompting", "vision-language"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "numpy>=1.23",
  "pillow>=9.0",
  "tqdm>=4.66",
  "torch>=2.0",
  "torchvision>=0.15",
]

[project.optional-dependencies]
llm = [
  "openai>=1.0.0",
]
plot = [
  "matplotlib>=3.7",
]
clip = [
  # Official CLIP package is typically installed from GitHub:
  # pip install 'openvocab-eval[clip]' will NOT install this automatically unless your environment allows git deps.
  "ftfy",
  "regex",
  "tqdm",
]

[project.scripts]
openvocab-eval = "openvocab_eval.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]
